command_class_lib: EqualsByAnyOtherName function is more flexible - when object is put into the same place so we can't assign to it the other name,  it is cloned and this clone is compared (possible error if it has links to other objects in the document) - otherwise, object's name, links in the tree and flags are backed up, assigned from the one we compare with and then compared and restored. Works much better, all links are resolved but again, treat with caution.
Attention: SaveAsInfoCommands to the same files are considired equal no matter the time,
because time is coded into the name of command and we use EqualsByAnyOtherName function
It's right behaviour in fact because if we didn't add any commands here and saved to the same file, then
why save at all?

streaming_class_lib: commented debug lines, they still can be of use
if we have errors with EqualsByAnyOtherName function

table_func_lib: -FourierSeries method is introduced, with some options
- LoadFromTextTable: when we have some sort of tabbed table and need to
take 2 columns from it. Needs to be more adjustable then now.
- Integral is now called Integrate and Integrate is now CalculateArea
- Average and RMSValue functions are now available.