BitStream_lib: bunch of bugs fixed. Instead of ReadBits it now has ReadBitsSigned and ReadBitsUnsigned.
expression_lib: in progress, now all the parse errors are ESyntaxError.
When parsing expression, error is not raised but fCorrect is set false and
reason is set in fLastErrorMsg.

If try to get value of incorrect expression, syntaxErr is raised.

Cyclic reference is detected and error is shown.

Some more math functions: sqrt and cos. Need MOAR!