Real1D: implemented cheating_inversion, it works SLOWER than our combined algorithm!
Complex1D: I thought that array of record would be faster to access then 2
independent arrays, in fact it is really bit faster, but strange thing,
direct copy of record (A:=B) it SLOWER than A.Re:=B.Re; A.Im:=B.Im. Has to
do with register variables.

Image_on_demand: we found our error and eliminated it, I hope. Deadlock
occured because LoadThread wanted to synchronize with main thread when it
was waiting for all threads to finish. Now instead of long WaitForSingleEvent
we make CheckSynchronize time to time. Yay!